be/requirements.txt
fastapi
uvicorn 
websocket-client
python-dotenv
rel 